Classwork 25/08/2023

+++ lab1/Grade.c
@@ -6,9 +6,9 @@
 
 int main() {
 
-    int a, b, c, sum;
+    float a, b, c, sum;
     printf("Please enter your grades for Maths, Literature and English: ");
-    scanf("%d %d %d", &a, &b, &c);
+    scanf("%f %f %f", &a, &b, &c);
     sum = a * 2 + b * 2 + c;
     if (sum >= 40) {
         printf("You have passed.");
